Some UDFs have been migrated to C language.
