+++ expected/udf_tg_verify_tables_extra.out
@@ -2051,6 +2051,59 @@ SELECT tg_execute_ddl('DROP TABLE udf_test_table_1', 'tsurugidb');
  DROP TABLE
 (1 row)
 
+-- Test case: 3-5-10
+--- Test setup: DDL of the Tsurugi
+SELECT tg_execute_ddl('
+  CREATE TABLE udf_test_type_table_1 (col BLOB)
+', 'tsurugidb');
+ tg_execute_ddl 
+----------------
+ CREATE TABLE
+(1 row)
+
+--- Test setup: DDL of the PostgreSQL
+CREATE FOREIGN TABLE udf_test_type_table_1 (col bytea) SERVER tsurugidb;
+--- Test
+SELECT tg_verify_tables('tg_schema', 'tsurugidb', 'public', 'detail', true);
+                 tg_verify_tables                 
+--------------------------------------------------
+ {                                               +
+     "verification": {                           +
+         "remote_schema": "tg_schema",           +
+         "server_name": "tsurugidb",             +
+         "local_schema": "public",               +
+         "mode": "detail",                       +
+         "tables_on_only_remote_schema": {       +
+             "count": 0,                         +
+             "list": []                          +
+         },                                      +
+         "foreign_tables_on_only_local_schema": {+
+             "count": 0,                         +
+             "list": []                          +
+         },                                      +
+         "tables_that_need_to_be_altered": {     +
+             "count": 1,                         +
+             "list": [                           +
+                 "udf_test_type_table_1"         +
+             ]                                   +
+         },                                      +
+         "available_foreign_table": {            +
+             "count": 0,                         +
+             "list": []                          +
+         }                                       +
+     }                                           +
+ }
+(1 row)
+
+--- Test teardown: DDL of the PostgreSQL
+DROP FOREIGN TABLE udf_test_type_table_1;
+--- Test teardown: DDL of the Tsurugi
+SELECT tg_execute_ddl('DROP TABLE udf_test_type_table_1', 'tsurugidb');
+ tg_execute_ddl 
+----------------
+ DROP TABLE
+(1 row)
+
 -- Test case: 8-1-1
 --- Test setup: DDL of the Tsurugi
 DO $$
