+++ include/fdw/tsurugi_utils.h
@@ -34,7 +34,14 @@ bool tg_create_cursor(ForeignScanState* node);
 bool tg_execute_foreign_scan(TgFdwForeignScanState *fsstate, TupleTableSlot *tupleSlot);
 bool tg_prepare_direct_modify(TgFdwDirectModifyState *dmstate);
 bool tg_execute_direct_modify(ForeignScanState *node);
-List *tg_execute_import_foreign_schema(ImportForeignSchemaStmt* stmt, Oid serverOid);
+bool tg_execute_import_foreign_schema(ImportForeignSchemaStmt* stmt, Oid serverOid,
+									  List** commands);
+bool tg_execute_ddl_statement(Oid server_oid, const char* ddl_statement, char** result);
+bool tg_execute_show_tables(Oid server_oid, const char* tg_schema, const char* tg_server,
+							const char* mode, bool detail, bool pretty, char** result);
+bool tg_execute_verify_tables(Oid server_oid, const char* tg_schema, const char* tg_server,
+							  Oid pg_schema_oid, const char* pg_schema, const char* mode,
+							  bool detail, bool pretty, char** result);
 #ifdef __cplusplus
 }
 #endif
