+++ scripts/install_deps.sh
@@ -11,6 +11,8 @@ TG_CMAKE_CXX_FLAGS_RELWITHDEBINFO="-O2 -g"
 _TSURUGI_FDW_DIR="${_SCRIPTS_DIR}/.."
 _OGAWAYAMA_DIR="${_THIRD_PARTY_DIR}/ogawayama"
 _METADATA_MANAGER_DIR="${_THIRD_PARTY_DIR}/metadata-manager"
+_TAKATORI_DIR="${_THIRD_PARTY_DIR}/takatori"
+_MESSAGE_MANAGER_DIR="${_THIRD_PARTY_DIR}/message-manager"
 
 if ldconfig -p | grep -F --quiet libmpdec++; then
   echo -e "\n[SKIPPED Install_mpdecimal]"
@@ -33,12 +35,20 @@ else
   rm -fr build-mpdecimal
 fi
 
+echo -e "\n[Install Takatori]"
+cd "${_TAKATORI_DIR}"
+rm -fr build
+mkdir -p build
+cd build
+cmake -G Ninja -DCMAKE_BUILD_TYPE=${TG_CMAKE_BUILD_TYPE} -DCMAKE_CXX_FLAGS_RELWITHDEBINFO="${TG_CMAKE_CXX_FLAGS_RELWITHDEBINFO}" -DBUILD_TESTS=OFF -DBUILD_DOCUMENTS=OFF -DCMAKE_PREFIX_PATH="${_FDW_DEPS_INSTALL_DIR}" -DCMAKE_INSTALL_PREFIX="${_FDW_DEPS_INSTALL_DIR}" ..
+cmake --build . --target install
+
 echo -e "\n[Install Ogawayama Stub]"
 cd ${_OGAWAYAMA_DIR}
 rm -fr build
 mkdir -p build
 cd build
-cmake -G Ninja -DCMAKE_BUILD_TYPE=${TG_CMAKE_BUILD_TYPE} -DCMAKE_CXX_FLAGS_RELWITHDEBINFO="${TG_CMAKE_CXX_FLAGS_RELWITHDEBINFO}" -DBUILD_TESTS=OFF -DBUILD_EXAMPLES=OFF -DSHARKSFIN_IMPLEMENTATION=shirakami -DCMAKE_PREFIX_PATH="${_FDW_DEPS_INSTALL_DIR}" -DCMAKE_INSTALL_PREFIX="${_FDW_DEPS_INSTALL_DIR}" -DCMAKE_MODULE_PATH="${_FDW_DEPS_INSTALL_DIR}" ..
+cmake -G Ninja -DCMAKE_BUILD_TYPE=${TG_CMAKE_BUILD_TYPE} -DCMAKE_CXX_FLAGS_RELWITHDEBINFO="${TG_CMAKE_CXX_FLAGS_RELWITHDEBINFO}" -DBUILD_TESTS=OFF -DBUILD_EXAMPLES=OFF -DSHARKSFIN_IMPLEMENTATION=shirakami -DCMAKE_PREFIX_PATH="${_FDW_DEPS_INSTALL_DIR}" -DCMAKE_INSTALL_PREFIX="${_FDW_DEPS_INSTALL_DIR}" -DCMAKE_MODULE_PATH="${_FDW_DEPS_INSTALL_DIR}" -DBUILD_STUB_ONLY=ON ..
 cmake --build . --target install
 
 echo -e "\n[Generate header file for protobuf]"
@@ -60,6 +70,7 @@ BUILDINFO="BUILD_TIMESTAMP:${BUILD_TIMESTAMP}
 ###### tsurugi_fdw and dependent modules
 - tsurugi_fdw $(git -C ${_TSURUGI_FDW_DIR} log --pretty="format:[%h - %s](${PROJECT_URL}/tsurugi_fdw/commit/%H)" -1 HEAD)
 - ogawayama $(git -C ${_OGAWAYAMA_DIR} log --pretty="format:[%h - %s](${PROJECT_URL}/ogawayama/commit/%H)" -1 HEAD)
+- takatori $(git -C ${_TAKATORI_DIR} log --pretty="format:[%h - %s](${PROJECT_URL}/takatori/commit/%H)" -1 HEAD)
 "
 echo "${BUILDINFO}" > "${_SCRIPTS_DIR}/FDW_BUILDINFO.md"
 echo "------------------------------------"
