+++ src/fdw/tsurugi_utils.cpp
@@ -22,6 +22,10 @@
 #include <boost/multiprecision/cpp_int.hpp>
 #include <boost/algorithm/string.hpp>
 #include <boost/format.hpp>
+
+#define BOOST_BIND_GLOBAL_PLACEHOLDERS
+#include <boost/property_tree/json_parser.hpp>
+
 #include <ogawayama/stub/api.h>
 #include <ogawayama/stub/error_code.h>
 
@@ -32,7 +36,10 @@
 extern "C" {
 #endif
 #include "postgres.h"
+#include "catalog/pg_type.h"
+#include "executor/spi.h"
 #include "utils/rel.h"
+
 #include "fdw/tsurugi_fdw.h"
 #ifdef __cplusplus
 }
@@ -497,29 +504,13 @@ tg_execute_direct_modify(ForeignScanState* node)
  *  Import Foreign Schema Functions.
  * 
  */
-List *
-tg_execute_import_foreign_schema(ImportForeignSchemaStmt* stmt, Oid serverOid)
+bool
+tg_execute_import_foreign_schema(ImportForeignSchemaStmt* stmt, Oid serverOid, List **commands)
 {
-    elog(DEBUG1, "tsurugi_fdw : %s", __func__);
+	elog(DEBUG1, "tsurugi_fdw : %s", __func__);
 
-    
 	ERROR_CODE error = ERROR_CODE::UNKNOWN;
-
-    /*
-	 * Checking the options of the Import Foreign Schema statement.
-	 * If the option is specified, an error is assumed.
-	 */
-	if ((stmt->options != nullptr) && (stmt->options->length > 0))
-	{
-#if PG_VERSION_NUM >= 130000
-		auto def = static_cast<DefElem*>(lfirst(stmt->options->elements));
-#else
-		auto def = static_cast<DefElem*>(lfirst(stmt->options->head));
-#endif
-		ereport(ERROR,
-				(errcode(ERRCODE_FDW_INVALID_OPTION_NAME),
-				 errmsg(R"(unsupported import foreign schema option "%.64s")", def->defname)));
-	}
+	*commands = nullptr;
 
 	/* Get information about foreign server and user mapping. */
 	ForeignServer* server = GetForeignServer(serverOid);
@@ -534,13 +525,8 @@ tg_execute_import_foreign_schema(ImportForeignSchemaStmt* stmt, Oid serverOid)
 	error = Tsurugi::tsurugi().get_list_tables(server->serverid, tg_table_list);
 	if (error != ERROR_CODE::OK)
 	{
-		/*
-		ereport(ERROR,
-			(errcode(ERRCODE_FDW_UNABLE_TO_CREATE_REPLY),
-			 errmsg("Failed to retrieve table list from Tsurugi. (error: %d)",
-				static_cast<int>(error)),
-			 errdetail("%s", Tsurugi::tsurugi().get_error_message(error).c_str())));
-		*/
+		Tsurugi::tsurugi().report_error("Failed to retrieve table list from Tsurugi.", error);
+		return false;
 	}
 
 	auto tg_table_names = tg_table_list->get_table_names();
@@ -587,7 +573,7 @@ tg_execute_import_foreign_schema(ImportForeignSchemaStmt* stmt, Oid serverOid)
 	}
 #endif	// ENABLE_IMPORT_TABLE_LIMITS
 
-	List* result_commands = NIL;
+	// List* result_commands = NIL;
 	/* CREATE FOREIGN TABLE statments */
 	for (const auto& table_name : tg_table_names)
 	{
@@ -596,12 +582,10 @@ tg_execute_import_foreign_schema(ImportForeignSchemaStmt* stmt, Oid serverOid)
 		/* Get table metadata from Tsurugi. */
 		error = Tsurugi::tsurugi().get_table_metadata(server->serverid, table_name, tg_table_metadata);
 		if (error != ERROR_CODE::OK)
-		{	/*
-			ereport(ERROR,
-				(errcode(ERRCODE_FDW_UNABLE_TO_CREATE_REPLY),
-				errmsg("Failed to retrieve table metadata from Tsurugi. (error: %d)",
-					static_cast<int>(error)),
-				errdetail("%s", Tsurugi::tsurugi().get_error_message(error).c_str())));	*/
+		{
+			Tsurugi::tsurugi().report_error(
+					"Failed to retrieve table metadata from Tsurugi.", error);
+			return false;
 		}	
 
 		/* Get table metadata from Tsurugi. */
@@ -616,15 +600,12 @@ tg_execute_import_foreign_schema(ImportForeignSchemaStmt* stmt, Oid serverOid)
 			/* Convert from Tsurugi datatype to PostgreSQL datatype. */
 			auto pg_type = tsurugi::convert_type_to_pg(column.atom_type());
 			if (!pg_type)
-			{	/*
-				ereport(ERROR,
-					(errcode(ERRCODE_FDW_INVALID_DATA_TYPE),
-					 errmsg(R"(unsupported tsurugi data type "%d")",
-					 	static_cast<int>(column.atom_type())),
-					 errdetail(R"(table:"%s" column:"%s" type:%d)",
-					 	table_name.c_str(),
-					 	column.name().c_str(),
-					 	static_cast<int>(column.atom_type()))));*/
+			{
+				auto err_mesg =
+					boost::format(R"(unsupported tsurugi data type "%d". (table:"%s" column:"%s"))") %
+					static_cast<int>(column.atom_type()) % table_name % column.name();
+				Tsurugi::tsurugi().report_error(err_mesg.str(), ERROR_CODE::UNSUPPORTED);
+				return false;
 			}
 			std::string type_name(*pg_type);
 
@@ -647,8 +628,503 @@ tg_execute_import_foreign_schema(ImportForeignSchemaStmt* stmt, Oid serverOid)
 
 		elog(DEBUG1, "%.512s", table_def.c_str());
 
-		result_commands = lappend(result_commands, pstrdup(table_def.c_str()));
+		*commands = lappend(*commands, pstrdup(table_def.c_str()));
 	}
 
-    return result_commands;
+    return true;
+}
+
+/**
+ * @brief Execute the DDL in Tsurugi.
+ * @param server_oid tsurugi server oid.
+ * @param ddl_statement ddl statement.
+ * @param result execution ddl command.
+ * @retval true success.
+ * @retval false failure.
+ */
+bool
+tg_execute_ddl_statement(Oid server_oid, const char* ddl_statement, char** result)
+{
+	ERROR_CODE error = ERROR_CODE::UNKNOWN;
+
+	error = Tsurugi::tsurugi().start_transaction(server_oid);
+	if (error != ERROR_CODE::OK) {
+		Tsurugi::tsurugi().report_error("Failed to start_transaction().", error, ddl_statement);
+		return false;
+	}
+
+	std::size_t num_rows;
+	error = Tsurugi::tsurugi().execute_statement(ddl_statement, num_rows);
+	if (error != ERROR_CODE::OK) {
+		Tsurugi::tsurugi().rollback();
+
+		Tsurugi::tsurugi().report_error("Failed to execute_statement().", error, ddl_statement);
+		return false;
+	}
+
+	error = Tsurugi::tsurugi().commit();
+	if (error != ERROR_CODE::OK) {
+		Tsurugi::tsurugi().report_error("Failed to commit().", error, ddl_statement);
+		return false;
+	}
+
+	/* Convert to uppercase for DDL command comparison. */
+	std::smatch match;
+	std::string ddl_command(ddl_statement);
+	std::regex_search(ddl_command, match, std::regex(R"_(^\s*([^\s]+)\s+([^\s]+))_"));
+
+	ddl_command = std::regex_replace(match.str(), std::regex(R"(^\s+|\s+$)"), "");
+	ddl_command = std::regex_replace(ddl_command, std::regex(R"(\s+)"), " ");
+	std::transform(ddl_command.begin(), ddl_command.end(), ddl_command.begin(),
+				   [](unsigned char c) { return std::toupper(c); });
+
+	*result = pstrdup(ddl_command.c_str());
+
+	return true;
+}
+
+/**
+ * @brief Get a list of Tsurugi tables.
+ * @param server_oid tsurugi server oid.
+ * @param tg_schema tsurugi schema name.
+ * @param tg_server tsurugi server name.
+ * @param mode report mode argument string.
+ * @param detail detail report flag.
+ * @param pretty pretty report flag.
+ * @param result execution result report.
+ * @retval true success.
+ * @retval false failure.
+ */
+bool
+tg_execute_show_tables(Oid server_oid,
+	                     const char* tg_schema,
+											 const char* tg_server,
+											 const char* mode,
+											 bool detail,
+											 bool pretty,
+											 char** result)
+{
+	static constexpr const char* const kKeyRootObject = "remote_schema";
+	static constexpr const char* const kKeyRemoteSchema = "remote_schema";
+	static constexpr const char* const kKeyServerName = "server_name";
+	static constexpr const char* const kKeyMode = "mode";
+	static constexpr const char* const kKeyRemoteTable = "tables_on_remote_schema";
+	static constexpr const char* const kKeyCount = "count";
+	static constexpr const char* const kKeyList = "list";
+
+	ERROR_CODE error;
+	TableListPtr tg_table_list;
+
+	/* Get a list of table names from Tsurugi. */
+	error = Tsurugi::tsurugi().get_list_tables(server_oid, tg_table_list);
+	if (error != ERROR_CODE::OK) {
+		Tsurugi::tsurugi().report_error("Failed to retrieve table list from Tsurugi.", error);
+		return false;
+	}
+
+	boost::property_tree::ptree table_list;  // table list array
+	/* Convert list of table names to ptree. */
+	for (const auto& table_name : tg_table_list->get_table_names()) {
+		boost::property_tree::ptree item;
+		item.put("", table_name);
+		table_list.push_back(std::make_pair("", item));
+	}
+
+	boost::property_tree::ptree pt_root;  // root object
+	boost::property_tree::ptree remote_schema;  // <remote_schema> object
+	boost::property_tree::ptree remote_tables;  // <tables_on_remote_schema> object
+
+  /* Add to table count. */
+	remote_tables.put(kKeyCount, table_list.size());
+
+	/* If the report level is 'detail', add a table listing. */
+	if (detail) {
+		/* Add to table name list. */
+		remote_tables.add_child(kKeyList, table_list);
+	}
+
+  /* Add to <remote_schema>. */
+	remote_schema.put(kKeyRemoteSchema, tg_schema);
+	/* Add to <server_name>. */
+	remote_schema.put(kKeyServerName, tg_server);
+	/* Add to <mode>. */
+	remote_schema.put(kKeyMode, mode);
+	/* Add to <tables_on_remote_schema> object. */
+	remote_schema.add_child(kKeyRemoteTable, remote_tables);
+
+  /* Add to root object. */
+	pt_root.add_child(kKeyRootObject, remote_schema);
+
+	std::stringstream ss;
+	/* Convert to JSON. */
+	try {
+		boost::property_tree::json_parser::write_json(ss, pt_root, pretty);
+	} catch (const std::exception& e) {
+		Tsurugi::tsurugi().report_error("JSON parser error", ERROR_CODE::UNKNOWN);
+		return false;
+	}
+	std::string json_str(ss.str());
+
+	/* Remove trailing newline code. */
+	if (json_str.back() == '\n') {
+		json_str.erase(json_str.size() - 1);
+	}
+
+	std::string separator = (pretty ? " " : "");
+
+	/* Converts the value of a numeric item from a string to a number. */
+	auto pattern_num = (boost::format(R"_("%s":\s*"(\d+)")_") % kKeyCount).str();
+	auto replace_num = (boost::format(R"("%s":%s$1)") % kKeyCount % separator).str();
+	json_str = std::regex_replace(json_str, std::regex(pattern_num), replace_num);
+
+	/* Converts an empty value of an array item from an empty character to an empty array. */
+	auto pattern_array = (boost::format(R"("%s":\s*"")") % kKeyList).str();
+	auto replace_array = (boost::format(R"("%s":%s[])") % kKeyList % separator).str();
+	json_str = std::regex_replace(json_str, std::regex(pattern_array), replace_array);
+
+	*result = pstrdup(json_str.c_str());
+
+	return true;
+}
+
+/**
+ * @brief Verify remote tables and local external tables.
+ * @param server_oid tsurugi server oid.
+ * @param tg_schema tsurugi schema name.
+ * @param tg_server tsurugi server name.
+ * @param pg_schema_oid postgresql schema oid.
+ * @param pg_schema postgresql schema name.
+ * @param mode report mode argument string.
+ * @param detail detail report flag.
+ * @param pretty pretty report flag.
+ * @param result execution result report.
+ * @retval true success.
+ * @retval false failure.
+ */
+bool
+tg_execute_verify_tables(Oid server_oid,
+	                     const char* tg_schema,
+											 const char* tg_server,
+											 Oid pg_schema_oid,
+											 const char* pg_schema,
+											 const char* mode,
+											 bool detail,
+											 bool pretty,
+											 char** result)
+{
+	static constexpr const char* const kKeyRootObject = "verification";
+	static constexpr const char* const kKeyRemoteSchema = "remote_schema";
+	static constexpr const char* const kKeyServerName = "server_name";
+	static constexpr const char* const kKeyLocalSchema = "local_schema";
+	static constexpr const char* const kKeyMode = "mode";
+	static constexpr const char* const kKeyRemoteOnly = "tables_on_only_remote_schema";
+	static constexpr const char* const kKeyLocalOnly = "foreign_tables_on_only_local_schema";
+	static constexpr const char* const kKeyAltered = "tables_that_need_to_be_altered";
+	static constexpr const char* const kKeyAvailable = "available_foreign_table";
+	static constexpr const char* const kKeyCount = "count";
+	static constexpr const char* const kKeyList = "list";
+	static const std::unordered_map<std::string, std::string> tz_abbreviate_type = {
+		{"time without time zone", "time"}, {"timestamp without time zone", "timestamp"}};
+
+	ERROR_CODE error = ERROR_CODE::UNKNOWN;
+	TableListPtr tg_table_list;
+
+	/* Get a list of table names from tsurugi. */
+	error = Tsurugi::tsurugi().get_list_tables(server_oid, tg_table_list);
+	if (error != ERROR_CODE::OK) {
+		Tsurugi::tsurugi().report_error("Failed to retrieve table list from Tsurugi.", error);
+		return false;
+	}
+
+	boost::property_tree::ptree list_remote;  // <tables_on_only_remote_schema> <list> array
+	boost::property_tree::ptree list_local;  // <foreign_tables_on_only_local_schema> <list> array
+	boost::property_tree::ptree list_altered;  // <tables_that_need_to_be_altered> <list> array
+	boost::property_tree::ptree list_available;  // <available_foreign_table> <list> array
+
+	if (SPI_connect() != SPI_OK_CONNECT) {
+		Tsurugi::tsurugi().report_error("Failed to SPI_connect.", ERROR_CODE::UNKNOWN);
+		return false;
+	}
+
+	static const int kValRelname = 1;
+	static const int kValAttname = 2;
+	static const int kValDatatype = 3;
+	static const int kValAtttypmod = 4;
+	static constexpr const char* const kMetadataQuery =
+		"SELECT "
+		"  c.relname, a.attname, format_type(a.atttypid, a.atttypmod) AS datatype, a.atttypmod "
+		"FROM pg_foreign_table ft "
+		"  JOIN pg_class c ON ft.ftrelid = c.oid "
+		"  JOIN pg_namespace n ON c.relnamespace = n.oid "
+		"  JOIN pg_attribute a ON a.attrelid = c.oid "
+		"WHERE "
+		"  (n.oid = $1) AND (ft.ftserver = $2) AND (a.attnum > 0) AND (NOT a.attisdropped) "
+		"ORDER BY c.relname, a.attnum";
+
+	Oid argtypes[2] = {OIDOID, OIDOID};
+	Datum values[2] = {ObjectIdGetDatum(pg_schema_oid), ObjectIdGetDatum(server_oid)};
+	char nulls[2] = {' ', ' '};
+
+	/* Refer to the system catalog. */
+	int res =
+		SPI_execute_with_args(kMetadataQuery, sizeof(values), argtypes, values, nulls, true, 0);
+	if (res != SPI_OK_SELECT) {
+		auto err_mesg = boost::format("Failed to SPI_execute_with_args. (error: %d)") % res;
+		Tsurugi::tsurugi().report_error(err_mesg.str(), ERROR_CODE::UNKNOWN);
+		return false;
+	}
+
+	// Metadata of foreign tables
+	std::unordered_map<std::string, std::vector<std::tuple<std::string, std::string, int, int>>>
+		foreign_table_metadata = {};
+
+	// List of Tsurugi table names
+	auto tsurugi_table_names = tg_table_list->get_table_names();
+
+	std::string skip_table_name = "";
+
+	/* Verifies whether foreign tables in the local schema exist in the remote schema,
+	 * and if so, retrieves metadata for the external tables.
+	 */
+	for (uint64 i = 0; i < SPI_processed; i++) {
+		HeapTuple spi_tuple = SPI_tuptable->vals[i];
+		TupleDesc tupdesc = SPI_tuptable->tupdesc;
+
+		// Foreign table name
+		std::string rel_name(SPI_getvalue(spi_tuple, tupdesc, kValRelname));
+
+		if (rel_name == skip_table_name) {
+			continue;
+		}
+
+		/* Add to a table that exists only in the remote schema. */
+		auto ite = std::find(tsurugi_table_names.begin(), tsurugi_table_names.end(), rel_name);
+		if (ite == tsurugi_table_names.end()) {
+			elog(DEBUG2, R"(Tables that do not exist in the remote schema. "%s")",
+				 rel_name.c_str());
+
+			/* Add to a table that exists only in the remote schema. */
+			list_local.push_back(std::make_pair("", boost::property_tree::ptree(rel_name)));
+
+			skip_table_name = rel_name;
+			continue;
+		}
+		skip_table_name = "";
+
+		// column name
+		char* column_name = SPI_getvalue(spi_tuple, tupdesc, kValAttname);
+
+		// data type (string)
+		std::string datatype(SPI_getvalue(spi_tuple, tupdesc, kValDatatype));
+		std::transform(datatype.begin(), datatype.end(), datatype.begin(), ::tolower);
+
+		bool is_null;
+		// precision
+		int precision = -1;
+		// scale
+		int scale = -1;
+
+		Datum typmod_datum = SPI_getbinval(spi_tuple, tupdesc, kValAtttypmod, &is_null);
+		int32 typmod = (!is_null ? DatumGetInt32(typmod_datum) : -1);
+		if (typmod != -1) {
+			precision = ((typmod - VARHDRSZ) >> 16) & 0xFFFF;
+			scale = (typmod - VARHDRSZ) & 0xFFFF;
+		}
+
+		/* Stores metadata for foreign tables. */
+		foreign_table_metadata[rel_name].emplace_back(std::make_tuple(column_name, datatype, precision, scale));
+	}
+
+	SPI_finish();
+
+	/* Verifies whether tables in the remote schema exist in the local schema. */
+	for (auto ite = tsurugi_table_names.begin(); ite != tsurugi_table_names.end();) {
+		/* Verify that a remote table exists on the local. */
+		if (foreign_table_metadata.find(*ite) == foreign_table_metadata.end()) {
+			elog(DEBUG2, R"(Tables that do not exist in the local schema. "%s")", (*ite).c_str());
+
+			/* Add to a table that exists only in the remote schema. */
+			list_remote.push_back(std::make_pair("", boost::property_tree::ptree(*ite)));
+			/* Exclude from metadata validation. */
+			tsurugi_table_names.erase(ite);
+
+			continue;
+		}
+		ite++;
+	}
+
+	/* Metadata Validation */
+	for (const auto& table_name : tsurugi_table_names) {
+		elog(DEBUG2, R"(Metadata Validation: table name: "%s")", table_name.c_str());
+
+		TableMetadataPtr tsurugi_table_metadata;
+		/* Get table metadata from Tsurugi. */
+		error = Tsurugi::tsurugi().get_table_metadata(server_oid, table_name, tsurugi_table_metadata);
+		if (error != ERROR_CODE::OK) {
+			Tsurugi::tsurugi().report_error("Failed to retrieve table metadata from tsurugi.",
+											error);
+			return false;
+		}
+
+		/* Get table metadata from PostgreSQL. */
+		const auto& pg_columns = foreign_table_metadata.find(table_name)->second;
+		/* Get table metadata from tsurugi. */
+		const auto& tg_columns = tsurugi_table_metadata->columns();
+
+		/* Validate the number of columns. */
+		if (pg_columns.size() != static_cast<size_t>(tg_columns.size())) {
+			elog(DEBUG2, "Number of columns does not match. local:%lu / remote:%d",
+					pg_columns.size(), tg_columns.size());
+
+			boost::property_tree::ptree item;
+			item.put("", table_name);
+			/* Add to a table with different column definitions. */
+			list_altered.push_back(std::make_pair("", item));
+
+			continue;
+		}
+
+		bool matched = true;
+		int idx = 0;
+		/* Validate the column metadata. */
+		for (const auto& pg_column : pg_columns) {
+			// Foreign table metadata
+			const auto& [pg_col_name, pg_col_type, pg_col_precision, pg_col_scale] = pg_column;
+			const auto& tg_column = tg_columns.at(idx++);
+
+			elog(DEBUG5, R"(Column Validation: column name: "%s")", pg_col_name.c_str());
+
+			/* Validate the column name. */
+			if (pg_col_name != tg_column.name()) {
+				elog(DEBUG2,
+					R"_(Name of column does not match. position:%d (local:"%s" / remote:"%s"))_",
+					idx, pg_col_name.c_str(), tg_column.name().c_str());
+
+				matched = false;
+				break;
+			}
+
+			/* Convert from tsurugi datatype to PostgreSQL datatype. */
+			auto remote_type_pg = tsurugi::convert_type_to_pg(tg_column.atom_type());
+			if (!remote_type_pg) {
+				elog(DEBUG2, "Data type is unknown. %s (atom_type:%d)", pg_col_name.c_str(),
+						static_cast<int>(tg_column.atom_type()));
+
+				matched = false;
+				break;
+			}
+
+			/* Correct time zone date/time type for PostgreSQL. */
+			std::vector<std::string> local_type = {pg_col_type};
+			const auto& pg_type = tz_abbreviate_type.find(pg_col_type);
+			if (pg_type != tz_abbreviate_type.end()) {
+				local_type.push_back(pg_type->second);
+			}
+
+			/* Validate the data type. */
+			auto ite = std::find(local_type.begin(), local_type.end(), *remote_type_pg);
+			if (ite == local_type.end()) {
+				elog(DEBUG2,
+						R"_(Datatype of column does not match. %s (local:"%s" / remote:"%s"))_",
+						pg_col_name.c_str(), local_type[0].c_str(), remote_type_pg->data());
+
+				matched = false;
+				break;
+			}
+
+			/* Validate the precision and scale. */
+			if ((pg_col_precision != -1) ||  (pg_col_scale != -1)) {
+				elog(
+					DEBUG2,
+					R"_(Column precision/scale does not match. "%s (local: %s(%d,%d) / remote:%s))_",
+					pg_col_name.c_str(), local_type[0].c_str(), pg_col_precision, pg_col_scale,
+					remote_type_pg->data());
+
+				matched = false;
+				break;
+			}
+		}
+
+		boost::property_tree::ptree list_item;
+		list_item.put("", table_name);
+		if (matched) {
+			/* Add to the matching table. */
+			list_available.push_back(std::make_pair("", list_item));
+		} else {
+			/* Add to a table with different column definitions. */
+			list_altered.push_back(std::make_pair("", list_item));
+		}
+	}
+
+	boost::property_tree::ptree pt_root;  // root object
+	boost::property_tree::ptree verification;  // <verification> object
+
+	/* Add to <remote_schema>. */
+	verification.put(kKeyRemoteSchema, tg_schema);
+	/* Add to <server_name>. */
+	verification.put(kKeyServerName, tg_server);
+	/* Add to <local_schema>. */
+	verification.put(kKeyLocalSchema, pg_schema);
+	/* Add to <mode>. */
+	verification.put(kKeyMode, mode);
+
+	/* Child object of a validation object. */
+	std::map<const char*, const boost::property_tree::ptree*> child_object = {
+		{kKeyRemoteOnly, &list_remote},
+		{kKeyLocalOnly, &list_local},
+		{kKeyAltered, &list_altered},
+		{kKeyAvailable, &list_available}};
+
+	/* Verification object is configured. */
+	for (const auto& object : child_object) {
+		const auto key = object.first;
+		const auto list = object.second;
+
+		boost::property_tree::ptree child;
+		/* Add to table count. */
+		child.put(kKeyCount, list->size());
+
+		/* If the report level is 'detail', add a table listing. */
+		if (detail) {
+			/* Add to table name list. */
+			child.add_child(kKeyList, *list);
+		}
+
+		/* Add to parent object. */
+		verification.add_child(key, child);
+	}
+
+	/* Add to root object. */
+	pt_root.add_child(kKeyRootObject, verification);
+
+	std::stringstream ss;
+	/* Convert to JSON. */
+	try {
+		boost::property_tree::json_parser::write_json(ss, pt_root, pretty);
+	} catch (const std::exception& e) {
+		Tsurugi::tsurugi().report_error("JSON parser error", ERROR_CODE::UNKNOWN);
+		return false;
+	}
+	std::string json_str(ss.str());
+
+	/* Remove trailing newline code. */
+	if (json_str.back() == '\n') {
+		json_str.erase(json_str.size() - 1);
+	}
+
+	std::string separator = (pretty ? " " : "");
+
+	/* Converts the value of a numeric item from a string to a number. */
+	auto pattern_num = (boost::format(R"_("%s":\s*"(\d+)")_") % kKeyCount).str();
+	auto replace_num = (boost::format(R"("%s":%s$1)") % kKeyCount % separator).str();
+	json_str = std::regex_replace(json_str, std::regex(pattern_num), replace_num);
+
+	/* Converts an empty value of an array item from an empty character to an empty array. */
+	auto pattern_array = (boost::format(R"("%s":\s*"")") % kKeyList).str();
+	auto replace_array = (boost::format(R"("%s":%s[])") % kKeyList % separator).str();
+	json_str = std::regex_replace(json_str, std::regex(pattern_array), replace_array);
+
+	*result = pstrdup(json_str.c_str());
+
+	return true;
 }
